refactor: 유틸성 클래스인 ProfitCalculator의 행위를 PlayerResult 객체에서 하도록 변경
