+++ src/main/kotlin/blackjack/domain/Cards.kt
@@ -9,9 +9,7 @@ class Cards(values: List<Card> = emptyList()) {
         _values.add(card)
     }
 
-    fun pick(): Card {
-        return _values.removeFirst()
-    }
+    fun pick(): Card = _values.removeFirst()
 
     fun getScore(): Int {
         val sum = values.sumOf { it.denomination.score }
